Restructures to export types for recurly-js
