+++ types/recurly__recurly-js/lib/address.ts
@@ -1,0 +1,12 @@
+export type Address = {
+  first_name: string;
+  last_name: string;
+  address1?: string;
+  address2?: string;
+  city?: string;
+  state?: string;
+  postal_code?: string;
+  country?: string;
+  phone?: string;
+  vat_number?: string;
+};
