+++ types/recurly__recurly-js/lib/emitter.ts
@@ -1,0 +1,9 @@
+type Listener = (...args: any[]) => void;
+export interface Emitter<Event = string> {
+  on(event: Event, listener: Listener): Emitter<Event>;
+  once(event: Event, listener: Listener): Emitter<Event>;
+  off(event?: Event, listener?: Listener): Emitter<Event>;
+  emit(event: Event, ...args: any[]): Emitter<Event>;
+  listeners(event: Event): Listener[];
+  hasListeners(event: Event): boolean;
+}
