+++ types/recurly__recurly-js/lib/error.ts
@@ -1,0 +1,6 @@
+export interface RecurlyError extends Error {
+  code: string;
+  message: string;
+  classification: string;
+  help?: string;
+}
