+++ types/recurly__recurly-js/lib/gift-card.ts
@@ -1,0 +1,14 @@
+import { RecurlyError } from './error';
+
+export type GiftCardOptions = {
+  code: string;
+};
+
+export type GiftCardResult = {
+  currency: string;
+  unit_amount: number;
+};
+
+export type Done = (error: RecurlyError, result: GiftCardResult) => void;
+
+export type GiftCard = (giftCardOptions: GiftCardOptions, done: Done) => void;
