+++ types/recurly__recurly-js/lib/validate.ts
@@ -1,0 +1,6 @@
+export type Validate = {
+  cardNumber: (cardNumber: string | number) => boolean;
+  cardType: (cardNumber: string | number, partial: boolean) => boolean;
+  expiry: (month: string | number, year: string | number) => boolean;
+  cvv: (cvv: string | number) => boolean;
+};
