+++ types/recurly__recurly-js/test/recurly-error.ts
@@ -1,0 +1,10 @@
+import { RecurlyError } from 'recurly__recurly-js';
+
+const recurlyError: RecurlyError = {
+  code: 'code',
+  classification: 'classification',
+  message: 'message',
+  name: 'name'
+};
+
+recurlyError.help = 'help';
