+++ types/recurly__recurly-js/tsconfig.json
@@ -28,6 +28,7 @@
         "test/elements.ts",
         "test/gift-card.ts",
         "test/paypal.ts",
+        "test/recurly-error.ts",
         "test/validate.ts",
         "test/pricing/checkout.ts",
         "test/pricing/subscription.ts",
